feat: some refactoring and IdCode tests
IdCode remains hard to parse. I think I might need to use dumpvars to switch
lexers again.
