+++ vcd/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     srcs = [
         "lexer.go",
         "parser.go",
+        "var_t.go",
     ],
     importpath = "github.com/filmil/go-vcd-parser/vcd",
     visibility = ["//visibility:public"],
@@ -23,7 +24,6 @@ go_test(
         "parser_test.go",
     ],
     embed = [":vcd"],
-    deps = ["@com_github_davecgh_go_spew//spew"],
 )
 
 go_test(
