+++ vcd/parser.go
@@ -1,8 +1,6 @@
 package vcd
 
 import (
-	"fmt"
-	"strings"
 	"time"
 
 	participle "github.com/alecthomas/participle/v2"
@@ -11,7 +9,8 @@ import (
 
 // VCDFile represents a parsed Value Change Dump.
 // The inline definition here, is based on the IEEE Std 1364-2001 Version C,
-// page 331.
+// page 331. Plus some extensions that don't seem described there, but
+// happen in realistic VCD files.
 type VCDFile struct {
 	Pos lexer.Position
 
@@ -34,88 +33,6 @@ type DeclarationCommandT struct {
 	Upscope        bool       `parser:"| @KwUpscope @KwEnd"`
 	//DeclarationKeyword DeclarationKeywordT `@@`
 	//CommandText        *string             `(@String) @KwEnd?`
-}
-
-// Capture implements custom capturing of tokens into VarT.
-func (self *VarT) Capture(tokens []string) error {
-	if self.p == nil {
-		p, err := participle.Build[IdT](participle.UseLookahead(3))
-		if err != nil {
-			return fmt.Errorf("could not build mini parser: %w", err)
-
-		}
-		self.p = p
-	}
-	// Parsing 6 tokens total.
-	//
-	//	1    2     3 4 5   6
-	//
-	// `$var logic 1 ! clk[foo][bar] $end`
-	for _, t := range tokens {
-		t = strings.TrimSpace(t)
-		if t == "" {
-			continue
-		}
-		self.tokenCount++
-		switch self.tokenCount {
-		case 1: // First token to be read.
-			if t != "$var" {
-				return fmt.Errorf("expected keyword: `$var`, got: %v", t)
-			}
-		case 2: // Type
-			self.VarType = t
-			if self.GetVarKind() == VarKindUnknown {
-				return fmt.Errorf("unknown var type: `%v`", t)
-			}
-		case 3:
-			if _, err := fmt.Sscanf(t, "%d", &self.Size); err != nil {
-				return fmt.Errorf("expected length, got: %w", err)
-			}
-		case 4:
-			self.Code = t // This can be anything. Just consume it.
-		case 5: // This is where it gets tricky.
-			self.tokenCount-- // Make sure we come back to handle this again, until $end.
-			if t == "$end" {  // Try to extract identifier now.
-				idString := strings.Join(self.varTokens, "")
-				id, err := self.p.ParseString("<idString>", idString)
-				if err != nil {
-					return fmt.Errorf("could not parse Id: `%v`: %w", idString, err)
-				}
-				self.Id = *id
-				self.varTokens = nil
-			} else {
-				// While not accummulated yet, continue adding.
-				self.varTokens = append(self.varTokens, t)
-			}
-		}
-		return nil
-	}
-	return nil
-}
-
-type VarT struct {
-	Pos        lexer.Position
-	tokenCount int
-	varTokens  []string // Accumulated tokens that refer to the signal variable. Can be many.
-	p          *participle.Parser[IdT]
-
-	Kw      bool
-	VarType string
-	Size    int
-	Code    string
-	Id      IdT
-	KwEnd   bool
-}
-
-type IdT struct {
-	Name    string  `parser:"@Ident"`
-	Indices []*IdxT `parser:"@@*"`
-}
-
-type IdxT struct {
-	Index    *int `parser:"(\"[\" @Int \"]\""`
-	MsbIndex *int `| "[" @Int `
-	LsbIndex *int `":" @Int "]")`
 }
 
 type VarTypeT struct {
@@ -257,7 +174,7 @@ func (self TimeUnit) Multiplier() float64 {
 type ScopeT struct {
 	Scope     bool       `@KwScope`
 	ScopeKind ScopeKindT `@@`
-	Id        string     `@Identifier @KwEnd`
+	Id        string     `@Ident @KwEnd`
 }
 
 type ScopeKindCode int
@@ -367,13 +284,13 @@ type ScalarValueChangeT struct {
 	Pos lexer.Position
 
 	Value          ValueT `@@`
-	IdentifierCode string `| @IdCode`
+	IdentifierCode string `@IdCode`
 }
 
 type ValueT struct {
 	Pos lexer.Position
 
-	Value string `@("0" | "1" | "x" | "X"| "z" | "Z")`
+	Value string `parser:"@(\"0\" | \"1\" | \"x\" | \"X\"| \"z\" | \"Z\")"`
 }
 
 type VectorValueChangeT struct {
@@ -397,9 +314,9 @@ type VectorValueChange3T struct {
 	IdentifierCode string `@IdCode`
 }
 
-func NewParser() *participle.Parser[VCDFile] {
+func NewParser[T any]() *participle.Parser[T] {
 	// Needs a lexer definition.
-	return participle.MustBuild[VCDFile](
+	return participle.MustBuild[T](
 		participle.Lexer(NewLexer()),
 		// For " variable[foo], variable[foo:bar]"
 		participle.UseLookahead(5),
